weather utils
